Corrected data ingestion and added a query model for user queries

+++ backend/main.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI, File, BackgroundTasks, UploadFile
 from fastapi.middleware.cors import CORSMiddleware
-from models import health, privacy
+from models import health, privacy, query
 from services import ai_service, privacy_service
 from database import db
 import uuid
@@ -38,28 +38,32 @@ async def summary_data(user_response: privacy.RequestUserInfo):
     else:
         return privacy.SummaryResponse(scrubbed_text = scrubbed_result.response, is_scrubbed = scrubbed_result.is_scrubbed, summary = "AI is currently offline")
 
-def process_document(file_path: str):
+def process_document(file_path: str, original_filename: str):
     text_chunks = []
     metadata = []
     uuids = []
     chunk_size, chunk_overlap = 1000, 200
-    file = fitz.open(file_path)
-    for page in file:
-        text = page.get_text() # extracting the text from each page
-        start_index = 0
-        while start_index < len(text):
-            text_chunks.append(text[start_index: start_index + chunk_size])
-            metadata.append({"source": file_path, "page": page.number})
-            uuids.append(uuid.uuid4().hex)
-            start_index += chunk_size - chunk_overlap
-    file.close()
-    os.remove(file_path)
-    if text_chunks:
-        db.collection.add(documents = text_chunks, metadatas = metadata, ids = uuids)
-
+    try:
+        with fitz.open(file_path) as file:
+            for page in file:
+                text = page.get_text() # extracting the text from each page
+                start_index = 0
+                while start_index < len(text):
+                    text_chunks.append(text[start_index: start_index + chunk_size])
+                    metadata.append({"source": original_filename, "page": page.number})
+                    uuids.append(uuid.uuid4().hex)
+                    start_index += chunk_size - chunk_overlap
+           
+        if text_chunks:
+            db.collection.add(documents = text_chunks, metadatas = metadata, ids = uuids)
+            print(f"Success! Shredded and saved {len(text_chunks)} chunks to ChromaDB.")
     
-
-
+    except Exception as e:
+        print(f"Background error: {e}")
+    finally:
+        if os.path.exists(file_path):
+            os.remove(file_path)
+    
 @app.post("/ingest")
 async def ingest_data(file: UploadFile, dispatcher: BackgroundTasks):
     unique_name =f"{uuid.uuid4().hex}.pdf"
@@ -67,9 +71,12 @@ async def ingest_data(file: UploadFile, dispatcher: BackgroundTasks):
     # opening the directory to which the file has to be temporarily uploaded in binary write mode
     with open(file_path, "wb") as buffer:
         shutil.copyfileobj(file.file, buffer)
-    dispatcher.add_task(process_document, file_path)
+    dispatcher.add_task(process_document, file_path, file.filename)
     return {"message": "File accepted. Processing in background."}
 
-
-
+@app.post("/query")
+async def user_query(request: query.Query):
+    question_text = request.QueryRequest
+    db_dict = db.retrieve_context(question_text)
+    return db_dict
